[feature/2] food repository, 식당 관련 api test code 추가 및 디버깅

+++ src/main/java/com/develop/datajpa/dto/baseball/ReviewDto.java
@@ -11,7 +11,7 @@ public class ReviewDto {
 
     private Long idx;
 
-    private Integer restaurantsId;
+    private Long restaurantsId;
 
     private Integer star;
