+++ src/main/java/com/develop/datajpa/repository/FoodRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Repository
 public interface FoodRepository extends JpaRepository<Food, Long> {
 
-    List<Food> findByRestaurantsId(int id);
+    List<Food> findByRestaurantsId(long id);
 
 }
